rename s3ObjectsFromStacks to getS3ObjectsFromStacks
