+++ lib/plugins/aws/deploy/lib/cleanupS3Bucket.js
@@ -3,7 +3,7 @@
 const BbPromise = require('bluebird');
 const _ = require('lodash');
 const findAndGroupStacks = require('../utils/findAndGroupStacks');
-const s3ObjectsFromStacks = require('../utils/s3ObjectsFromStacks');
+const getS3ObjectsFromStacks = require('../utils/getS3ObjectsFromStacks');
 
 module.exports = {
   getObjectsToRemove() {
@@ -24,7 +24,7 @@ module.exports = {
         const stacks = findAndGroupStacks(response, service, stage);
         const stacksToKeep = _.takeRight(stacks, stacksToKeepCount);
         const stacksToRemove = _.pullAllWith(stacks, stacksToKeep, _.isEqual);
-        const objectsToRemove = s3ObjectsFromStacks(stacksToRemove, service, stage);
+        const objectsToRemove = getS3ObjectsFromStacks(stacksToRemove, service, stage);
 
         if (objectsToRemove.length) {
           return BbPromise.resolve(objectsToRemove);
