+++ lib/plugins/aws/deploy/tests/all.js
@@ -10,4 +10,4 @@ require('./updateStack');
 require('./index');
 require('./configureStack');
 require('./findAndGroupStacks');
-require('./s3ObjectsFromStacks');
+require('./getS3ObjectsFromStacks');
