+++ lib/plugins/aws/deploy/tests/getS3ObjectsFromStacks.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const expect = require('chai').expect;
-const s3ObjectsFromStacks = require('../utils/s3ObjectsFromStacks');
+const getS3ObjectsFromStacks = require('../utils/getS3ObjectsFromStacks');
 
-describe('#s3ObjectsFromStacks()', () => {
+describe('#getS3ObjectsFromStacks()', () => {
   it('should return an empty result in case no stacks are provided', () => {
-    expect(s3ObjectsFromStacks([], 'test', 'dev')).to.deep.equal([]);
+    expect(getS3ObjectsFromStacks([], 'test', 'dev')).to.deep.equal([]);
   });
 
   it('should return an empty result in case no stacks are provided', () => {
@@ -41,6 +41,6 @@ describe('#s3ObjectsFromStacks()', () => {
       { Key: 'serverless/test/dev/1476779278222-2016-10-18T08:27:58.222Z/test.zip' },
     ];
 
-    expect(s3ObjectsFromStacks(stacks, 'test', 'dev')).to.deep.equal(expected);
+    expect(getS3ObjectsFromStacks(stacks, 'test', 'dev')).to.deep.equal(expected);
   });
 });
