+++ lib/plugins/aws/deploy/utils/getS3ObjectsFromStacks.js
